Working on fixing API/implementing basic ORM
